connectors/sink/azure-cosmosdb/requirements.txt
python-dotenv==1.0.1
glassflow
azure-cosmos==4.6.0